fix somes bug backend + car + car details fe

+++ EcommerWebCar/main-service/src/main/java/com/tlaq/main_service/dto/responses/carResponse/CarServiceResponse.java
@@ -3,6 +3,8 @@ package com.tlaq.main_service.dto.responses.carResponse;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,13 +14,13 @@ public class CarServiceResponse {
     Long id;
     String dongCo;
     String hopSo;
-    Double congSuat;
-    Double momenXoan;
-    Double dungTichXiLanh;
-    Double dungTichXang;
-    Double taiTrong;
-    Double chieuDai;
+    BigDecimal congSuat;
+    BigDecimal momenXoan;
+    BigDecimal dungTichXiLanh;
+    BigDecimal dungTichXang;
+    BigDecimal taiTrong;
+    BigDecimal chieuDai;
     String mauSac;
-    Double tocDoToiDa;
+    BigDecimal tocDoToiDa;
     String loaiNhienLieu;
 }
